[스프링 - 협업 미션] 찰리(박민영) 미션 제출합니다.

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,29 +1,45 @@
 package wooteco.subway.station.application;
 
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
+import wooteco.subway.exception.badrequest.DuplicateUniqueKeyException;
+import wooteco.subway.exception.badrequest.IllegalDeleteException;
+import wooteco.subway.exception.notfound.StationNotFoundException;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.dto.StationTransferLinesDto;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
-        Station station = stationDao.insert(stationRequest.toStation());
-        return StationResponse.of(station);
+        try {
+            Station station = stationDao.insert(stationRequest.toStation());
+            return StationResponse.of(station);
+        } catch (DuplicateKeyException e) {
+            throw new DuplicateUniqueKeyException("이미 존재하는 지하철 역입니다");
+        }
     }
 
     public Station findStationById(Long id) {
-        return stationDao.findById(id);
+        return stationDao.findById(id).orElseThrow(StationNotFoundException::new);
+    }
+
+    public List<StationTransferLinesDto> findAllWithTransferLines() {
+        return stationDao.findAllWithTransferLines();
     }
 
     public List<StationResponse> findAllStationResponses() {
@@ -34,7 +50,21 @@ public class StationService {
                 .collect(Collectors.toList());
     }
 
+    public StationResponse updateStation(Long id, StationRequest stationRequest) {
+        try {
+            Station station = stationDao.findById(id).orElseThrow(StationNotFoundException::new);
+            Station updateStation = station.update(stationRequest.getName());
+            stationDao.update(updateStation);
+            return StationResponse.of(updateStation);
+        } catch (DuplicateKeyException e) {
+            throw new DuplicateUniqueKeyException("이미 존재하는 지하철 역입니다");
+        }
+    }
+
     public void deleteStationById(Long id) {
+        if (sectionDao.existStation(id)) {
+            throw new IllegalDeleteException("이미 노선에 등록된 지하철 역입니다");
+        }
         stationDao.deleteById(id);
     }
 }
